move rewrite into DecomposeConcat.cpp
