+++ compiler/src/iree/compiler/GlobalOptimization/BUILD.bazel
@@ -48,7 +48,7 @@ iree_compiler_cc_library(
         "CleanupNumericNarrowing.cpp",
         "Convert1X1FilterConv2DToMatmul.cpp",
         "DataLayoutPropagation.cpp",
-        "DecomposeConcat.cpp",
+        "DecomposeTensorOps.cpp",
         "DemoteContractionInputsToBF16.cpp",
         "DetachElementwiseFromNamedOps.cpp",
         "EraseUnusedLinalgOperands.cpp",
