+++ compiler/src/iree/compiler/GlobalOptimization/CMakeLists.txt
@@ -44,7 +44,7 @@ iree_cc_library(
     "CleanupNumericNarrowing.cpp"
     "Convert1X1FilterConv2DToMatmul.cpp"
     "DataLayoutPropagation.cpp"
-    "DecomposeConcat.cpp"
+    "DecomposeTensorOps.cpp"
     "DemoteContractionInputsToBF16.cpp"
     "DetachElementwiseFromNamedOps.cpp"
     "EraseUnusedLinalgOperands.cpp"
