+++ compiler/src/iree/compiler/GlobalOptimization/Passes.td
@@ -21,12 +21,12 @@ def Convert1X1FilterConv2DToMatmul:
   let constructor = "mlir::iree_compiler::GlobalOptimization::createConvert1X1FilterConv2DToMatmulPass()";
 }
 
-def DecomposeConcat :
-    Pass<"iree-global-opt-decompose-concat", ""> {
-  let summary = "Decomposes concatenations into a destination and a sequence of slice inserts.";
-  let constructor = "mlir::iree_compiler::GlobalOptimization::createDecomposeConcatPass()";
+def DecomposeTensorOps:
+    Pass<"iree-global-opt-decompose-tensor-ops", ""> {
+  let summary = "Decomposes extractions/concatenations into a destination and a sequence of slice inserts.";
+  let constructor = "mlir::iree_compiler::GlobalOptimization::createDecomposeTensorOpsPass()";
   let options = [
-    Option<"enableConcatTransposition", "enable-concat-transposition", "bool",
+    Option<"enableTransposition", "enable-transposition", "bool",
            /*default=*/"false", "Allows transposing concatenations such that "
                                 "they occur on the inner most dims.">,
   ];
